Move sum, count to Array refinement methods
