+++ lib/stupidedi.rb
@@ -16,7 +16,6 @@ require "ruby/object"
 require "ruby/module"
 require "ruby/array"
 require "ruby/hash"
-require "ruby/enumerable"
 require "ruby/string"
 require "ruby/blank"
 require "ruby/to_d"
