+++ lib/stupidedi/editor/004010.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     #
