+++ lib/stupidedi/editor/004010/N4.rb
@@ -1,4 +1,6 @@
 class Stupidedi
+  using Refinements
+
   module Editor
 
     class FortyTenEd
