+++ lib/stupidedi/editor/X12.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     class X12 < AbstractEd
