+++ lib/stupidedi/editor/X222-HC837.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     class X222 < AbstractEd
