+++ lib/stupidedi/editor/claim_ack.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     # Generates a 277CA acknowledgement
