+++ lib/stupidedi/editor/interchange_ack.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     # Generates a TA1 acknowledgment
