+++ lib/stupidedi/editor/result_set.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     class ResultSet
