+++ lib/stupidedi/editor/transaction_set_ed.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     class TransactionSetEd < AbstractEd
