+++ lib/stupidedi/editor/transmission_ed.rb
@@ -1,4 +1,6 @@
 module Stupidedi
+  using Refinements
+
   module Editor
 
     class TransmissionEd < AbstractEd
